Move Padlock into the catalogue app
